Removed compilation warnings.
git-svn-id: file:///fltk/svn/branches/branch-3.0@9333 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ include/fltk3/TextEditor.h
@@ -40,7 +40,7 @@
 namespace fltk3 {
 
   // key will match in any state
-  const unsigned int TEXT_EDITOR_ANY_STATE = -1L;
+  const unsigned int TEXT_EDITOR_ANY_STATE = (unsigned)(-1);
   
   /**
    This is the FLTK text editor widget. It allows the user to
